Renamed Docker Int Driver to Linux Int Driver
The Docker Integration driver has been renamed to the more appropriate
Linux Integration driver. All of the functionality is the same except
the driver's key is now "linux" instead of "docker" and the package is
"github.com/codedellemc/libstorage/drivers/integration/linux".
