--- imports/config/imports_config_11_int_docker.go
@@ -1,7 +1,0 @@
-// +build libstorage_integration_driver_docker
-
-package config
-
-func init() {
-	defaultIntDriver = "docker"
-}
