+++ imports/config/imports_config_11_int_linux.go
@@ -1,0 +1,7 @@
+// +build libstorage_integration_driver_linux
+
+package config
+
+func init() {
+	defaultIntDriver = "linux"
+}
