--- imports/local/imports_local_int_docker.go
@@ -1,8 +1,0 @@
-// +build libstorage_integration_driver_docker
-
-package local
-
-import (
-	// load the packages
-	_ "github.com/codedellemc/libstorage/drivers/integration/docker"
-)
