+++ imports/local/imports_local_int_linux.go
@@ -1,0 +1,8 @@
+// +build libstorage_integration_driver_linux
+
+package local
+
+import (
+	// load the packages
+	_ "github.com/codedellemc/libstorage/drivers/integration/linux"
+)
